requirements.txt
poe-api = 0.3.1
whisper = 1.1.10
PyAudio = 0.2.11
SpeechRecognition = 3.10.0
